pyproject.toml
[project]
name = "taxonomy-rag"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.1",
    "psycopg_pool>=3.1",
    "pgvector>=0.3",
    "pydantic-settings>=2.0",
    "sentence-transformers>=3.0",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langchain-ollama>=0.3",
    "langchain-community>=0.3",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pymupdf>=1.24",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "jupyter>=1.1", "ipykernel>=6.29"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taxonomy_rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
